requirements.txt
python-telegram-bot==20.7
requests==2.31.0
pandas==2.1.4
numpy==1.26.2
ta==0.11.0
scikit-learn==1.3.2
apscheduler==3.10.4